remove sqlserver
merge generic sql code into snowflake, its only consumer
