+++ flow/connectors/utils/peers.go
@@ -45,12 +45,6 @@ func CreatePeerNoValidate(
 			return wrongConfigResponse, nil
 		}
 		innerConfig = bqConfigObject.BigqueryConfig
-	case protos.DBType_SQLSERVER:
-		sqlServerConfigObject, ok := config.(*protos.Peer_SqlserverConfig)
-		if !ok {
-			return wrongConfigResponse, nil
-		}
-		innerConfig = sqlServerConfigObject.SqlserverConfig
 	case protos.DBType_S3:
 		s3ConfigObject, ok := config.(*protos.Peer_S3Config)
 		if !ok {
